dsl_auto_type::auto_type

Module expression_type_inference

Structs

InferrerSettings

InferrerSettingsBuilder

Functions

infer_expression_type

This is meant to be used if there’s need to infer a single expression type out of the context of a function. It will be assumed that there are no intermediate variables (let statements). initially, but one can still use intermediate block expression to annotate types.